backend/requirements.txt
# FavBox Backend Dependencies

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0        # SQLite async driver
asyncpg==0.30.0          # PostgreSQL async driver

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.2.0

# WebSocket
websockets==14.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.17

# Development
httpx==0.28.1            # For testing
